HUD w/ health bar

+++ SuperFishAl/Assets/Scripts/EnemyCollisionDamageComponent.cs
@@ -4,7 +4,7 @@ using System.Collections;
 
 public class EnemyCollisionDamageComponent : MonoBehaviour
 {
-    public double damageAmount = 0.5;
+    public float damageAmount = 0.5f;
     public float timeBetweenAttacks = 2f;
 
     private HealthComponent playerHealth;
